feat(provider): add canonical gemini package
