+++ .changeset/config.json
@@ -13,6 +13,7 @@
       "@robota-sdk/agent-provider-openai",
       "@robota-sdk/agent-provider-openai-compatible",
       "@robota-sdk/agent-provider-gemma",
+      "@robota-sdk/agent-provider-gemini",
       "@robota-sdk/agent-provider-google",
       "@robota-sdk/agent-provider-bytedance",
       "@robota-sdk/agent-transport-http",
