+++ .changeset/gemini-provider-package-rename.md
@@ -1,0 +1,7 @@
+---
+'@robota-sdk/agent-provider-gemini': minor
+'@robota-sdk/agent-provider-google': patch
+'@robota-sdk/agent-cli': patch
+---
+
+Add canonical Gemini provider package and keep the Google provider package as a compatibility wrapper.
