+++ packages/agent-cli/package.json
@@ -53,7 +53,7 @@
     "@robota-sdk/agent-core": "workspace:*",
     "@robota-sdk/agent-provider-anthropic": "workspace:*",
     "@robota-sdk/agent-provider-gemma": "workspace:*",
-    "@robota-sdk/agent-provider-google": "workspace:*",
+    "@robota-sdk/agent-provider-gemini": "workspace:*",
     "@robota-sdk/agent-provider-openai": "workspace:*",
     "@robota-sdk/agent-provider-qwen": "workspace:*",
     "@robota-sdk/agent-sdk": "workspace:*",
