+++ packages/agent-cli/src/utils/provider-default-definitions.ts
@@ -1,6 +1,6 @@
 import { createAnthropicProviderDefinition } from '@robota-sdk/agent-provider-anthropic';
 import { createGemmaProviderDefinition } from '@robota-sdk/agent-provider-gemma';
-import { createGeminiProviderDefinition } from '@robota-sdk/agent-provider-google';
+import { createGeminiProviderDefinition } from '@robota-sdk/agent-provider-gemini';
 import { createOpenAIProviderDefinition } from '@robota-sdk/agent-provider-openai';
 import { createQwenProviderDefinition } from '@robota-sdk/agent-provider-qwen';
 import type { IProviderDefinition } from './provider-definition.js';
