+++ packages/agent-provider-gemini/docs/README.md
@@ -1,0 +1,14 @@
+# Gemini Provider Docs Index
+
+## Scope
+
+- Documentation entry for `@robota-sdk/agent-provider-gemini`.
+
+## Canonical Documents
+
+- `SPEC.md`: Provider scope, ownership boundaries, and canonical responsibilities.
+
+## Notes
+
+- Keep provider-specific details in this package docs.
+- Keep shared contract-level rules in `@robota-sdk/agent-core` docs.
