+++ packages/agent-provider-gemini/examples/scenarios/gemini-image-dry-run.record.json
@@ -1,0 +1,23 @@
+{
+  "schemaVersion": 1,
+  "recordType": "scenario-record",
+  "scope": "packages/agent-provider-gemini",
+  "packageName": "@robota-sdk/agent-provider-gemini",
+  "command": {
+    "executable": "pnpm",
+    "args": ["scenario:verify"],
+    "rendered": "pnpm scenario:verify"
+  },
+  "cwd": "<cwd>",
+  "status": 0,
+  "stdout": {
+    "raw": "\n> @robota-sdk/agent-provider-gemini@3.0.0-beta.55 scenario:verify /Users/jungyoun/Documents/dev/robota/packages/agent-provider-gemini\n> pnpm --filter @robota-sdk/agent-core build && DRY_RUN=1 GEMINI_IMAGE_MODEL=gemini-2.0-flash-exp-image-generation pnpm exec tsx examples/example-generate-image.ts && DRY_RUN=1 GEMINI_IMAGE_MODEL=gemini-2.0-flash-exp-image-generation pnpm exec tsx examples/example-compose-image.ts\n\n\n> @robota-sdk/agent-core@3.0.0-beta.55 build /Users/jungyoun/Documents/dev/robota/packages/agent-core\n> tsup\n\n\u001b[34mCLI\u001b[39m Building entry: src/index.ts\n\u001b[34mCLI\u001b[39m Using tsconfig: tsconfig.json\n\u001b[34mCLI\u001b[39m tsup v8.5.0\n\u001b[34mCLI\u001b[39m Using tsup config: /Users/jungyoun/Documents/dev/robota/packages/agent-core/tsup.config.ts\n\u001b[34mCLI\u001b[39m Building entry: src/index.ts\n\u001b[34mCLI\u001b[39m Using tsconfig: tsconfig.json\n\u001b[34mCLI\u001b[39m tsup v8.5.0\n\u001b[34mCLI\u001b[39m Using tsup config: /Users/jungyoun/Documents/dev/robota/packages/agent-core/tsup.config.ts\n\u001b[34mCLI\u001b[39m Target: es2022\n\u001b[34mCLI\u001b[39m Target: es2022\n\u001b[34mCLI\u001b[39m Cleaning output folder\n\u001b[34mESM\u001b[39m Build start\n\u001b[34mCJS\u001b[39m Build start\n\u001b[34mCLI\u001b[39m Cleaning output folder\n\u001b[34mESM\u001b[39m Build start\n\u001b[34mDTS\u001b[39m Build start\n\u001b[34mDTS\u001b[39m Build start\n\u001b[32mESM\u001b[39m \u001b[1mdist/node/index.js \u001b[22m\u001b[32m130.19 KB\u001b[39m\n\u001b[32mESM\u001b[39m ⚡️ Build success in 235ms\n\u001b[32mCJS\u001b[39m \u001b[1mdist/node/index.cjs \u001b[22m\u001b[32m130.89 KB\u001b[39m\n\u001b[32mCJS\u001b[39m ⚡️ Build success in 235ms\n\u001b[32mESM\u001b[39m \u001b[1mdist/browser/index.js \u001b[22m\u001b[32m130.19 KB\u001b[39m\n\u001b[32mESM\u001b[39m ⚡️ Build success in 234ms\n\u001b[32mDTS\u001b[39m ⚡️ Build success in 811ms\n\u001b[32mDTS\u001b[39m \u001b[1mdist/browser/index.d.ts \u001b[22m\u001b[32m151.31 KB\u001b[39m\n\u001b[32mDTS\u001b[39m ⚡️ Build success in 834ms\n\u001b[32mDTS\u001b[39m \u001b[1mdist/node/index.d.ts  \u001b[22m\u001b[32m151.31 KB\u001b[39m\n\u001b[32mDTS\u001b[39m \u001b[1mdist/node/index.d.cts \u001b[22m\u001b[32m151.31 KB\u001b[39m\nDRY_RUN enabled. model=gemini-2.0-flash-exp-image-generation, promptLength=70\nDRY_RUN enabled. compose example configuration validated.\n",
+    "normalized": "DRY_RUN enabled. model=gemini-2.0-flash-exp-image-generation, promptLength=70\nDRY_RUN enabled. compose example configuration validated.",
+    "sha256": "6c740a3d54d90a645853d88e214d0210fcc61a78709c9e9cc5deaaa2ca35047f"
+  },
+  "stderr": {
+    "raw": "",
+    "normalized": "",
+    "sha256": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
+  }
+}
