+++ packages/agent-provider-gemini/src/image-operations.test.ts
