+++ packages/agent-provider-gemini/src/image-operations.ts
