+++ packages/agent-provider-gemini/src/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @robota-sdk/agent-provider-gemini package
+ *
+ * Provides a Gemini API provider implementation with provider-agnostic TUniversalMessage.
+ */
+
+// Main exports
+export * from './provider';
+export * from './provider-definition';
+export * from './types';
