+++ packages/agent-provider-gemini/src/message-converter.ts
