+++ packages/agent-provider-gemini/src/tool-schema-converter.ts
