+++ packages/agent-provider-gemini/src/types/api-types.ts
