+++ packages/agent-provider-gemma/README.md
@@ -2,7 +2,7 @@
 
 Gemma model-family provider for Robota using OpenAI-compatible local endpoints such as LM Studio.
 
-This provider is separate from Gemini API support. Gemini API behavior remains in `agent-provider-google` until the separate Gemini rename migration is completed.
+This provider is separate from Gemini API support. Gemini API behavior belongs to `agent-provider-gemini`; `agent-provider-google` remains only as a compatibility wrapper.
 
 The provider owns Gemma/LM Studio serving-template projection. It filters Gemma reasoning-channel markers from user-visible text and converts documented native tool-call text emitted by the Gemma/LM Studio template into Robota universal `toolCalls` when the referenced tool was declared in the request.
