+++ packages/agent-provider-google/docs/README.md
@@ -1,14 +1,3 @@
-# Google Provider Docs Index
+# Google Provider Compatibility Docs
 
-## Scope
-
-- Documentation entry for `@robota-sdk/agent-provider-google`.
-
-## Canonical Documents
-
-- `SPEC.md`: Provider scope, ownership boundaries, and canonical responsibilities.
-
-## Notes
-
-- Keep provider-specific details in this package docs.
-- Keep shared contract-level rules in `@robota-sdk/agent-core` docs.
+- [SPEC.md](SPEC.md) - compatibility package contract and migration path.
